use `SmallVec` for `Topic` type
