+++ xmtp_api_d14n/src/protocol/in_memory_cursor_store.rs
@@ -196,7 +196,7 @@ mod tests {
     }
 
     fn topic(name: &str) -> Topic {
-        Topic::from_bytes(name.as_bytes().to_vec())
+        Topic::from_bytes(name.as_bytes())
     }
 
     #[xmtp_common::test]
